Fix mypy strict errors in app (no type: ignore)
Resolve 4 no-any-return errors flagged by `mypy app` under strict mode,
using explicit types/casts rather than Any or ignores:
- topic_extraction_service: annotate _text_for(source: Source) and
  _normalized_candidates(extracted: list[ExtractedTopic]); removes two
  type: ignore comments and the resulting no-any-return
- core/logging.get_logger: cast structlog.get_logger(...) (typed Any) to
  structlog.stdlib.BoundLogger
- trend_score_calculator: wrap float ** float in float() (typeshed types
  it as Any)
- hackernews http_client: cast item_response.json() to dict[str, Any]

mypy: success (88 files); pytest: 24 passed; ruff: clean.

+++ services/api/app/application/use_cases/trends/trend_score_calculator.py
@@ -75,4 +75,5 @@ class TrendScoreCalculator:
         age_hours = (ctx.reference_time - latest).total_seconds() / 3600.0
         if age_hours <= 0:
             return 1.0
-        return 0.5 ** (age_hours / ctx.recency_half_life_hours)
+        # `float ** float` is typed as Any in typeshed; make the float explicit.
+        return float(0.5 ** (age_hours / ctx.recency_half_life_hours))
